fix: 필터 객체 쿼리 파라미터로 변환 로직 모듈화
