+++ src/hooks/useMembersQuery.ts
@@ -5,7 +5,7 @@ import dayjs from 'dayjs';
 
 import { getMembers, RawMember } from '@/apis/adminMembers';
 import { FilterType } from '@/types/filter.type';
-
+import { filtersToParams } from '@/utils/memberParams';
 const PAGE_SIZE = 8;
 
 /**
@@ -68,31 +68,8 @@ function getSex(sex: RawMember['sex']) {
  * @since 2025.05.13
  * @author 권민지
  */
-function useMembersQueryParams(filters: FilterType, page: number, size: number = PAGE_SIZE) {
-  return useMemo(() => {
-    const params: Record<string, string> = {};
-
-    if (filters.name) params.name = filters.name;
-    if (filters.sex && filters.sex !== 'ALL') params.sex = filters.sex;
-    if (filters.birthDateRange && filters.birthDateRange[0] && filters.birthDateRange[1]) {
-      params.birthDateStart = filters.birthDateRange[0];
-      params.birthDateEnd = filters.birthDateRange[1];
-    }
-    if (filters.memberStatus && filters.memberStatus !== 'ALL')
-      params.memberStatus = filters.memberStatus;
-    if (filters.createdDateRange && filters.createdDateRange[0] && filters.createdDateRange[1]) {
-      params.startDate = filters.createdDateRange[0];
-      params.endDate = filters.createdDateRange[1];
-    }
-    if (filters.hasLoan && filters.hasLoan !== 'ALL') params.hasLoan = filters.hasLoan;
-    if (filters.transactionCountMin)
-      params.transactionCountMin = String(filters.transactionCountMin);
-    if (filters.transactionCountMax)
-      params.transactionCountMax = String(filters.transactionCountMax);
-    params.page = String(page - 1);
-    params.size = String(size);
-    return params;
-  }, [filters, page, size]);
+function useMembersQueryParams(filters: FilterType, page: number, size: number = 8) {
+  return useMemo(() => filtersToParams(filters, page, size), [filters, page, size]);
 }
 
 /**
